Split the microcode into separate files and updated toolchain to support this.
The microcode was up to 650 lines and only getting longer, and organizing it within the one file was getting to be a pain. Now it's split into separate files per functional area, and should be easier to manage. The microcode generator tool now creates a merged file that the assembler and any other tools can use, so no changes are needed there.