Implemented the basics of multi-byte instructions and converted add16.ab to the new way.
Instructions can now be two bytes, using IR for the first byte and RR (the register register) for the second byte, and we can drive XY using an indirect reference via RR, like we've always been able to via IR. This is how instructions that use multiple registers will be implemented.

Converted add16.ab from a proof-of-concept with fixed registers to the new flexible way. The assembler doesn't really support this yet, so right now you have to put in add16.ab, followed by a 'data byte' of 0-3 to specify which register.