Fixed bug where x in test lines meant 'ignore'.
This was half-intentional but the side effect is that there was no way to test for a real 'x', and if there was an 'x' in the output the checks was ignored entirely. Some load instructions never worked right and are now failing the tests as they should, will fix those next.