Fixed bug in store where one byte of reg was ending up in both RAM bytes
We had one unused cycle during the jump from a/b/c/d version to ?
version, so I've filled that with the extra steps required. st16 is kind
of a mess now so should get optimized and cleaned up at some point, but
that will be a later point.

Also fixed the tests so they'll catch this kind of thing now!